LayoutTests/platform/ios-simulator/svg/zoom/page/zoom-zoom-coords-expected.txt

Test
Test
Test
This test checks getBoundingClientRect() on zoomed HTML and SVG elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Gray rectangles: 100x50
Green rectangles: 100x50, zoom=200%
Blue rectangles: 200x100, zoom=50%

Checking HTML elements:

PASS div1.left.toFixed(2) is "0.00"
PASS div1.top.toFixed(2) is "0.00"
PASS div1.width.toFixed(2) is "100.00"
PASS div1.height.toFixed(2) is "50.00"
PASS div1.right.toFixed(2) is "100.00"
PASS div1.bottom.toFixed(2) is "50.00"

PASS div2.left.toFixed(2) is "0.00"
PASS div2.top.toFixed(2) is "25.00"
PASS div2.width.toFixed(2) is "100.00"
PASS div2.height.toFixed(2) is "50.00"
PASS div2.right.toFixed(2) is "100.00"
PASS div2.bottom.toFixed(2) is "75.00"

PASS div3.left.toFixed(2) is "0.00"
PASS div3.top.toFixed(2) is "300.00"
PASS div3.width.toFixed(2) is "200.00"
PASS div3.height.toFixed(2) is "100.00"
PASS div3.right.toFixed(2) is "200.00"
PASS div3.bottom.toFixed(2) is "400.00"

Checking SVG elements:

PASS svg1.left.toFixed(2) is "0.00"
PASS svg1.top.toFixed(2) is "250.00"
PASS svg1.width.toFixed(2) is "150.00"
PASS svg1.height.toFixed(2) is "50.00"
PASS svg1.right.toFixed(2) is "150.00"
PASS svg1.bottom.toFixed(2) is "300.00"
PASS rect1.left.toFixed(2) is "0.00"
PASS rect1.top.toFixed(2) is "250.00"
PASS rect1.width.toFixed(2) is "100.00"
PASS rect1.height.toFixed(2) is "50.00"
PASS rect1.right.toFixed(2) is "100.00"
PASS rect1.bottom.toFixed(2) is "300.00"
PASS image1.left.toFixed(2) is "100.00"
PASS image1.top.toFixed(2) is "250.00"
PASS image1.width.toFixed(2) is "50.00"
PASS image1.height.toFixed(2) is "25.00"
PASS image1.right.toFixed(2) is "150.00"
PASS image1.bottom.toFixed(2) is "275.00"
PASS text1.left.toFixed(2) is "100.00"
PASS text1.top.toFixed(2) is "282.66"
FAIL text1.width.toFixed(2) should be 49.20. Was 47.44.
PASS text1.height.toFixed(2) is "6.16"
FAIL text1.right.toFixed(2) should be 149.20. Was 147.44.
PASS text1.bottom.toFixed(2) is "288.82"

PASS svg2.left.toFixed(2) is "75.00"
PASS svg2.top.toFixed(2) is "100.00"
PASS svg2.width.toFixed(2) is "150.00"
PASS svg2.height.toFixed(2) is "50.00"
PASS svg2.right.toFixed(2) is "225.00"
PASS svg2.bottom.toFixed(2) is "150.00"
PASS rect2.left.toFixed(2) is "75.00"
PASS rect2.top.toFixed(2) is "100.00"
PASS rect2.width.toFixed(2) is "100.00"
PASS rect2.height.toFixed(2) is "50.00"
PASS rect2.right.toFixed(2) is "175.00"
PASS rect2.bottom.toFixed(2) is "150.00"
PASS image2.left.toFixed(2) is "175.00"
PASS image2.top.toFixed(2) is "100.00"
PASS image2.width.toFixed(2) is "50.00"
PASS image2.height.toFixed(2) is "25.00"
PASS image2.right.toFixed(2) is "225.00"
PASS image2.bottom.toFixed(2) is "125.00"
PASS text2.left.toFixed(2) is "175.00"
PASS text2.top.toFixed(2) is "132.66"
FAIL text2.width.toFixed(2) should be 47.44. Was 48.33.
FAIL text2.height.toFixed(2) should be 5.94. Was 6.05.
FAIL text2.right.toFixed(2) should be 222.44. Was 223.33.
FAIL text2.bottom.toFixed(2) should be 138.60. Was 138.71.

PASS svg3.left.toFixed(2) is "900.00"
PASS svg3.top.toFixed(2) is "500.00"
PASS svg3.width.toFixed(2) is "300.00"
PASS svg3.height.toFixed(2) is "100.00"
PASS svg3.right.toFixed(2) is "1200.00"
PASS svg3.bottom.toFixed(2) is "600.00"
PASS rect3.left.toFixed(2) is "900.00"
PASS rect3.top.toFixed(2) is "500.00"
PASS rect3.width.toFixed(2) is "200.00"
PASS rect3.height.toFixed(2) is "100.00"
PASS rect3.right.toFixed(2) is "1100.00"
PASS rect3.bottom.toFixed(2) is "600.00"
PASS image3.left.toFixed(2) is "1100.00"
PASS image3.top.toFixed(2) is "500.00"
PASS image3.width.toFixed(2) is "100.00"
PASS image3.height.toFixed(2) is "50.00"
PASS image3.right.toFixed(2) is "1200.00"
PASS image3.bottom.toFixed(2) is "550.00"
PASS text3.left.toFixed(2) is "1100.00"
PASS text3.top.toFixed(2) is "565.33"
FAIL text3.width.toFixed(2) should be 98.41. Was 94.87.
PASS text3.height.toFixed(2) is "12.31"
FAIL text3.right.toFixed(2) should be 1198.41. Was 1194.88.
PASS text3.bottom.toFixed(2) is "577.64"

PASS successfullyParsed is true

TEST COMPLETE
